requirements.txt
Django==1.6.2
argparse==1.2.1
beautifulsoup4==4.3.2
dj-database-url==0.3.0
dj-static==0.0.5
django-debug-toolbar==1.0.1
django-toolbelt==0.0.1
gunicorn==18.0
httplib2==0.9
psycopg2==2.5.2
pystache==0.5.3
sqlparse==0.1.11
static==1.0.2
wsgiref==0.1.2
